+++ examples/ptx_runner/sm90/matmul-with-tma-v3.py
@@ -5,10 +5,10 @@ import torch
 DEVICE = triton.runtime.driver.active.get_active_torch_device()
 
 
-import triton_ml_runner
+import triton_runner
 
 # @triton.jit
-@triton_ml_runner.jit
+@triton_runner.jit
 def matmul_kernel_make_tensor_desciptor(a_ptr, b_ptr, c_ptr,  #
                                         M, N, K,  #
                                         BLOCK_SIZE_M: tl.constexpr, BLOCK_SIZE_N: tl.constexpr,
@@ -70,7 +70,7 @@ def matmul(a, b):
         BLOCK_SIZE_M=128,
         BLOCK_SIZE_K=64,
         BLOCK_SIZE_N=64,
-        ptx_dir=triton_ml_runner.get_file_dir(__file__)
+        ptx_dir=triton_runner.get_file_dir(__file__)
     )
 
     return c
