+++ setup.py
@@ -1,7 +1,7 @@
 import setuptools
 
 setuptools.setup(
-    name="triton_ml_runner",
+    name="triton_runner",
     version="0.0.1",
     author="Bob Huang",
     author_email="git@bobhuang.xyz",
